requirements.txt
Django==2.0.6
django-crispy-forms==1.7.2
django-filter==1.1.0
django-filters==0.2.1
djangorestframework==3.8.2
Pillow==5.1.0
python-decouple==3.1
pytz==2018.4
sorl-thumbnail==12.4.1
